I recovered mausezahn documentation and tar balls from archive.org and removed the scripts and ads from the pages. So it should look as "original" as possible.
The project rattenzahn wants to improve mausezahn as a stand alone tool
which is compact and efficient and a good tool to be run on plastic
routers like those supported by openwrt.
So one can use an old plastic router, laptop or cheap embedded internet of shit device as a packet generator to do network measurements.

Mausezahn has some flaws such as not sending packets if the payload file is
too big or if it is not existent. Without payload file mz will segfault.
At least this is what I expected while using mausezahn for network tests
floding a router with udp traffic in oder to find corrupted packets generated
by an overloaded router.
Also the performance should be improved as there are routers that contain
test generators with better flooding performance.

Especially the shell "mops" which is accessible using telnet of mausezahn
is very interesting and is something that needs to be improved and extended.

Hopefully rattenzahn will be the better mausezahn.

Levush, initial commit